requirements.txt
requests
discord.py[voice]
discord-py-slash-command
asyncio